Cleanup livestatus usage from RBN processing
In case a user configured the contact selection option "Members of
contact groups" (contact_groups), a Livestatus query was made to get the
members of the configured contact groups. This was done every time such
a rule matched a notification.

In specific situations such a lookup could cause dead locks between the
Microcore and the notify helper. Since the notification helpers have the
whole base configuration in memory, we can simply lookup the group
memberships without the Livestatus lookup.

Change-Id: Ia78b26153ea53ea39e3e7e4be5922e93eab268d8