Ensure livestatus lookup is never done with Microcore
As described in the comments, this code path is there to handle a Nagios
core specific situation and should never be reached with the Microcore.
Since it might be a source for a dead lock with the Microcore we now
really ensure that this is never reached with the Microcore.

Change-Id: Ie1f64fc60d46849ab1e4f9e92d40585dd386bbc7