Make UserId subclass of str with validation
UserId is used on several occasions to construct paths. A malicious
UserId could therefore lead to path traversal.
Since UserId is already a NewType we subclass str now and include
validation. Since we used empty UserIds in the past (which are not valid
usernames...) there is also the possibility to circumvent the validation
but only for empty UserIds.

Change-Id: Iae00e1317b2f8db671db5ce41fc078768e7fcb87