14288 FIX Fix Apache error message in case site Apache is not started
With 2.1.0p7 and werk #14281 we changed the system Apache configuration which
broke the System Apache error page that is shown to the user in case the site
Apache is not running. It showed a 'Service Unavailable' message instead of the
intended 'Checkmk: Site Not Started'.

To apply this fix, you will have to execute <tt>omd update-apache-config [site]</tt>
as root user after executing <tt>omd update</tt>.

Change-Id: If13a0323885ae8993e6a655f53c4badc3071033d