Use setproctitle package instead of broken home-grown code.
Setting the process title is a bit tricky and not very portable, setting the
commandline arguments is even trickier. For the gory details, look e.g.
into the kernel code, the prctl(2) man page or just at the article at
https://hackmag.com/coding/reverse-shell-237-bytes/.

Our own code was unportable and didn't excatly do what it intended to, in
addition it triggered some mypy issues which were real bugs (e.g.  str vs
bytes). So let's nuke that and build on the blood, sweat and tears of
others...

Change-Id: Ie0862620f180a814e0b65cefc0ac9f6e0835601a