Fix int and gui tests
Since ff9d0bfddd5fcfc89ebeeafbd00f78da32c24a37 we do not update/set the
cookie at every page load. Thanks to the redirect cascade the final
response in this test does not contain the cookie anymore. But the requests
session has it. Since the sole purpose is to check for a valid login and
the right cookie format we now test the cookiejar of the session.

Change-Id: I77ee20956153ce7c0320cc1333e2676b120fb8f7