wsgi middleware: typing
We need to import "_typesched.wsgi" because this is the type that
allows us to safely interface with external libraries using the same
type. mypy doesn't understand when you define the exact same type in
separate locations. For mypy cmk.gui.wsg.type_defs.WSGIApplication
is different to _typesched.wsgi.WSGIApplication, even though it is
a verbatim copy-paste.

Change-Id: I7dd1e903b7220c75ff861242b047ee3e7856a2dc