transform_json: use NotRequired
PEP 589 specified inheritance as the way to achieve a mix of optional
and mandatory keys in a TypedDict. However, with python 3.11 a newer
syntax is becoming available, which avoids defining multiple TypedDicts'
and therefore makes it possible to keys belonging to one definition. We
can access this new syntax via typing_extensions.

Change-Id: Ie3a754c9fb65d48d59fd4af37f521b4e0e56fa3b