Fix GUI crawl: Make feature toggle explicit
The listing of rules on the rule set page has an optional feature:
Analyse matching of the rules, visualized by the colored circles). This
mode was previously enabled once the `self._hostname` was set. However,
during the recent changes this was confused multiple times which seems
to be a good sign that we have to clean this up somehow.

Now this change tries to make it clearer by making the feature toggle
more explicit.

Change-Id: I5a82cb046b02e6b58d74dd259dfd13ac6b34a42d