14636 Extension package manager: Change in GUI workflows
This is sqash of the following commits:

1st: rename 'remove' and 'dis/enable'

Unfortunately we need the names to avoid collision and confusion.
The term "block" is not meant to stay, the functionality will be
consolidated in the end.

2nd: split up upload & installation

3rd: 14636 Extension package manager: Change in GUI workflows

4th: always create new MKPs

5th: create enabled packages during update config

6th: consolidate old and new disabeling concepts

7th: update installed packages

8th: Update plugins before loading config

A vanished ruleset result in all configured rules for that ruleset to be dropped irreversably, so it is vital to update the plugins first.

9th: fix typo

10th: fix typo

11th: improve error message

12th: fix utils path

13th: actually call update actions

Change-Id: Ifc1c339ceb26623e9ebe8998fbb2d9af62edcbb5