Remove hidden keyword argument from TextInput
This is a ValueSpec API change.

If you used the hidden keyword argument for TextInput or one of it's
subclasses ID, UUID, UserID, RegExp, EmailAddress, HostAddress, Url,
TextAreaUnicode, Filename or LDAPDistinguishedName please change your
code to use the Password ValueSpec.

The argument was also removed from Password and PasswordSpec, it had no
affect on those classes, so you can safely remove the argument.

Change-Id: I4b5cc066dc9d74a2268260991680e4c51f98051c