Unbreak lasso build
The 'make' command is executing python during processing and relied on
the OS Python. The build now failed because the implicit dependency on
six was removed from virtualenv which was needed here.

Since we have our own python in OMD, we should use this to have a
more deterministic build.

Change-Id: Ie382208c743a9e7374d63a5e4eb37a3ad234652c