14490 Limit number of checkers to number of available CPUs
When the global setting "Maximum concurrent Checkmk checkers" (topic "Monitoring core") is set to
a value greater than the number of available CPUs, the number of actual checkers is limited to the
number of CPUs. This is done because the checker processes are most effective when each checker
gets a dedicated CPU.
When the number of checkers is limited, a warning is displayed in the UI upon an activation of
pending changes.

Change-Id: Ic8e5a2b69ffdb7146104d43acf70940a2a84bb5e