agent-ctl: Replace sync channel with async
Motivation: This channel is to communicate with mailslot
server thread. We can't afford to synchronous wait for message
during 30-120 seconds. Alternatively we could implement processing
queue, but it may be even more complicated/

API is implemented, but not enabled(can't  be without proper
integration)

Tests: manual with failing unit test conditions

CMK-11076

Change-Id: I3316ba456f185d23c4de73e9116526a20fecbc12