14167 Improved loading times of user configured graphs/views/dashboards/reports
Configurations with hundreds or thousands of custom visualization elements such as custom graphs suffered
from slow page rendering, because the configuration files usually become quite large.
To improve the loading time, an attempt is now always made to load a cached version of the file in question.
The cached version is located in the tmpfs directory and also uses an optimized file format to reduce the parsing time by a
factor of 10.

Change-Id: Ibeacbf4488d9cf138556d5bad8146eac488afbaa