Extract pagetype instances
Previously the collection of instances of a pagetype was held by a class
member Pagetype.__instances which made it hard to understand the
relation and responsibilities of the involved classes and instances. Since
the collection was a class member, this was some sort of global state
which made it hard to understand when the collection was initialized and
when it could reliably be used.

The instances are now still loaded by CustomGraphPage.load which now
returns a OverridableInstances object. These objects are now propagated
as explicit dependency through the code, which makes the relations much
clearer.

Some instance collection related methods have now only been moved
between the classes and were not renamed yet, which will be one of the
next steps. Additionally there are some more methods which may also be
moved to OverridableInstances.

Change-Id: I43594996d21d27331f63d1da54ea025038ff5760