Added a workaround for broken black packacking.
The black package *always* installs blackd, even when it is not required via
the extras, see https://github.com/psf/black/issues/690.

This in turn leads to broken dependencies for blackd, which one of our
integration tests (test_python_modules) stumbles over by accident.

As a workaround, just add the "d" extra, even though we don't need it.  The
alternative of ignoring blackd in the test looks fragile, we would still
ship a broken blackd.

Change-Id: Id9355df6adf3c25840fe4b2fb4f8b484207b4925