14361 FIX Windows agent uses retry_count correctly
Previously, Windows agent used retry_count as a maximal allowed
count of plugin's failures. If the count of failures exceeded the
limit set by retry_count, then the plugin had been excluded from
execution. Such behavior may break monitoring or updating in
some cases.

Since this release, the retry_count is defined as a maximally
allowed count of attempts to call a plugin before clearing the
previously gathered data (also known as a cache). This approach
is fully compliant with the documentation, the help and the
functionality of the Linux agent.

The incompatibility in this werk is limited by definition: the
Windows agent will continue to call a plugin even if the number
of errors exceeds retry_count. This is the same behavior as when
retry_count is not set at all.

IMPORTANT: Windows agent always ignores retry_count for synchronous
plugins. Windows agent may decrease value of retry_count if the
value is set too high.

Change-Id: Ia1109fd10a5e4b90efd6a1fc894e3c99fb1fde46