14693 FIX clustering: Fix clustering in case of failover
In case of failover, a clustered check showed "Item not found in monitoring data"
instead of fetching the data from the second node.
The problem was that we didn't fetch a clustered service data on the nodes on
which the service was never discovered.

This behavior was visible when the rule "Aggregation options for clustered services"
was set on "Failover".

CMK-11126

Change-Id: Iefc6acfc4fd0134b1395594145831914fc2f0024