14721 FIX Crash using the cmk_site_statistics (core PIDs) filter
When using the cmk_site_statistics (core PIDs) filter with no monitoring data present the
view would crash.

Note: This filter is not intended to be used in views and should only be used in host and service
problem graphs.
But due to compatibility the filter is not removed from view filters.

Now the view no longer crashes but rather displays a warning.

Change-Id: Ib82ca30120e02ab3ac15aaa6ef4758de057d9cc4