14694 FIX bluecoat_sensors: Fix duplicate service names
This change is incompatible for users who use bluecoat_sensors_temp check.

During the check plugin migration, bluecoat_sensors_temp service name was
changed. This caused a clash with the bluecoat_sensors service name.
The bluecoat_sensors_temp disappeared and there was a warning on activate
changes.

After the upgrade, run the discovery on the affected hosts to get the
bluecoat_sensors_temp service again.

SUP-11277

Change-Id: I773a8f2742055e95bf3098fa3eb2d0774f622480