14493 Add global setting "Enable community translations"
* Defaults to False - per default only English and German are shown in
the language dropdown (User > Edit profile > Language)
* Changing this global setting requires an additional activation of
pending changes
* When changed from True to False, the UI default and user specific
languages are checked and set to English if a community translation was
in place. Any such reset to English is logged as a warning to the
details column on the "Activate pending changes" page.
* When updating from an older version without this global setting, the
UI default and user specific languages are checked and if a community
translation was in place this global setting is changed to True to keep
the previous language configuration.

CMK-11138

Change-Id: I327f637307e8519c20c1a2e1ed76d552c18de364