Prevent duplicate dual-stack PING service (CRE)
Consider having a dual-stacked host with IPv4 preferred. Previously, the
"PING IPv6" service was written to the Nagios config unconditionally,
even when you already had a service by that name for that host
(presumably manually configured).

With this patch, the automatic check is only written to the config file
if there is no other check with that description already. This also
enables the user to tune the parameters of the service.