14552 FIX Invalid rule configuration for static checks after upgrade
When upgrading a 1.6.0 site, the configuration data for manual checks
(also known as enforced services or static checks) would only be
partially updated.  Specifically, the checktype would not be updated to
reflect new plug-in names.  This would result in warnings such as this,

-| WARNING: Invalid rule configuration detected (Ruleset: static_checks:fileinfo-groups, Title: Size, age and count of file groups, Folder: ,

-| Rule nr: 1, Exception: fileinfo.groups is not an existing element in this selection.)

while upgrading from 2.0.0 to 2.1.0. With this werk, the checktype is
automatically transformed during the update step.

SUP-11280

Change-Id: Iaaf396c9e5bd7bbe613455142177426e023d4705