Remove thousand_sep for Float and Integer
This is a ValueSpec API change.

Before this change there was a thousand_sep argument to Float.__init__
and Intger.__init__ which was not used at all.
Setting thousand_sep for Float to a value other than None would lead to
wrong value:

vs.Float(thousand_sep=",", unit="unit").value_to_html(1002003) == "1,002,003,.00 unit"

As this is hard to fix, and thousand_sep is not used at all it was
removed instead of fixed.

Change-Id: I19f75f84a585eb8d2855c0a9a74ee4a82234fc3e