Prepare dashlet config for typing
With this change we prepare for making DashletConfig a typed dict. We
introduce a dashlet type specific dashlet config class which defines the
attribute that can be configured for that dashlet while keeping
DashletConfig a regular dict. This way we can prepare the code step py
step for the final change.

Change-Id: I6462076814df63371602a6ad45bc2412ba1db2dc