14735 FIX Agent Bakery: Cleanup orphaned packages
When narrowing down packages using the "Agent bakery packages" agent ruleset,
it may happen that a host is assigned to two different agent packages after baking.

This may happen when you start using the abovementioned rule. On baking, this will result
in a new agent package for the matched hosts. If nothing else happens to the package
that the hosts were assigned to previously, the now-unselected platforms will remain
for the previous package.

As they are explicitly unwanted by setting the rule, these orphaned packages are now cleaned up.

Change-Id: Ib189d034223eee58cccac9db7e1ed975b8082471