move 'version' endpoint to tag group 'Monitoring'
If the agent controller tries to register a host via agent labels, the receiver will write a request for registration to disk.
Since it will only be further processed if the site is a CPE, we want to fail early if it is not (otherwise the registration would just not terminate).

For this we need this API endpoint to be accessible even if the 'Setup' tag group is not available, which is the case for a distributed setup.

Change-Id: I7fcefb6c47e94da7aeddd489546f19161500a356