make version endpoint available for 'Monitoring'
The API endpoint is used by the agent receiver to check whether it is OK to write a request for registration to disk (which is only the case for CPE).
This needs to work even for remote sites with disabled WATO, where endpoints of the 'Setup' tag group are not available.

Change-Id: I7fcefb6c47e94da7aeddd489546f19161500a356