14293 FIX Enforce use of new apache hooks before updating to 2.2
With #14281 a new system apache configuration hook for managing the system
apache reverse proxy config has been introduced.

In all previous versions the transition was highly recommended for security
reasons, but not enforced to stay compatible. As of Checkmk 2.2 the new
mechanism will be enforced.

If you try to update an existing site to 2.2 which has not been switched to the
new mechanism with <tt>omd update-apache-config [SITE]</tt>, then the update
will be aborted.

Change-Id: If692c210a2fa33d1e93d31f284b565e5313c3b12