Make build-cmk-packages script behave differently based on Testbuild role
With the new JJB generated Jenkins jobs we don't configure distros as
folder properties anymore and it's hard/impossible to give a job parameter
different default values based on the actual job. To still let the
Testbuild behave as before editions are now saved in editions.json for
all cases and using the OMD package cache get's disabled when the
script is used in Testbuild context.

 #jjb-migration

Change-Id: Ibd735151b43a43fbca278ede7a420f23a2f2d7e5