activate changes unit tests: clarify paths that are actually expected
Due to our entangled code structure (see comments), it was unclear what
paths are actually expected for the site synchronisation vs. what has
just been added as a workaround to deal with the issue of paths being
registered simply because the edition-specific code is present in the
repository.

It is now clear what is actually expected.

Also, a bug is fixed due to the functions
is_{raw,enterprise,managed}_edition() not working properly with the test
set-up.

Change-Id: I00f1402ff130daec1d224d86e3b8f7437718295c