Figure dashlet ajax call: Use real dashlet object
Instead of using some arbitrary fake dashlet object we hand over all
necessary information from the caller to the AJAX endpoint to be able to
construct the real dashlet object (dashboard name and dashlet it).

The JS code still needs some information from the dashlet configuration.
To make this possible we now explicitly hand over the dashlet
configuration to the JS code. Also makes things a little clearer to hand
over it explicitly instead of extracting it from the post body.

This is a preparation on the way to move dashboard configs to TypedDict.

Change-Id: Iadd9dfdd19623281a9b736e80babc3f15960726d