fix creation mode when getting parameters for edit page
In a previous refactoring we set the page mode to "edit" even when we
were in "create" mode and this led to some errors, for example when
creating a new rule for "Graph Tuning", the graph type dropdown was
supposed to be readonly only in edit mode but it was like that in
create mode as well.

This commit fixes the issue by passeing the correct mode to the function
that creates the ValueSpec for the page.

Change-Id: I2a793a62c753834c6c59e4e311ec502fbf6f84aa