14561 FIX Prometheus: allow operation of special agent without querying targets endpoint
Previously, the Prometheus special agent would query the endpoint
`/api/v1/targets` regardless of what was configured in the Prometheus
rule. This was an additional burden on users, which did not specify any
scrape targets and therefore do not need this endpoint. E.g., it can
lead to additional timeouts. With this werk, the endpoint
`/api/v1/targets` is only queried if necessary.

CMK-11288

Change-Id: I5a0bed07f97b1217a8b2d632dd5abe82527082ea