Change DashboardConfig to TypedDict
This change migrates the dashboard configurations to the TypedDict
DashboardConfig. With this change the dashboard is now the first of 3
visual types to be a typed dict.

Since we have a lot of shipped dashboard configurations this commit is
fairly big. However, most of the changes are updated lines caused by
wrapping the data structures in the TypedDicts, removing useless
attributes and adding some missing attributes to the definitions.

From now on every builtin dashboard configuration is fully type checked.

Change-Id: Ifb7f78103e4b79fc451119f63e526eaaf796cbf3