Autouse ip lookup fixture
Some tests might require to lookup IP adresses, but
by default, they use actual networking features of
the os. This might and did lead to problems because
the underlying environment is outside the control
of the test.
This change uses the up lookup fixture by default in
all python tests.

Change-Id: I71c2104844ccc4249206c820ee9d197156e34a2a