Flatten RulesetCollection spaghetti
I guess that makes the class hierarchy into linguine, now.

Anyway, the call stack is clearer when it is explicit.

This is one of the many cases where we have an overly complex (and
useless) inheritance hierarchy.

Moreover, the collection now looks a bit more like a container and less
magic Ruleset instantiation happens within it.

We will need that to decouple the bakery from the Ruleset code.

CMK-11212

Change-Id: I4a100a03e2d68c4c5a53e421e1fbcfc8fa45cc6b