14739 FIX Regression: Activate agent controller daemon on Debian based systems
This is a regression that occurs with Checkmk 2.1.0p11.

When installing the prepackaged <tt>check-mk-agent_2.1.0p11-1_all.deb</tt>
agent package on a Linux host, the installation scripts miss to activate
the agent controller daemon, that is responsible for providing the agent port.<br>
Agent bakery packages (including the VANILLA package) are not affected.

You can activate it manually by calling the scripts
<tt>/var/lib/cmk-agent/scripts/cmk-agent-useradd.sh</tt> and
<tt>/var/lib/cmk-agent/scripts/super-server/setup trigger</tt>
(in that order) on affected hosts after agent package installation.

Change-Id: I603e8cdb719a93bc7704aa232a2c83c76bd9354c