Prepare element config for typing
With this change we start to prepare for making ElementConfig a typed
dict. We now subsequently introduce a specific config and parameter
class which defines the attribute that can be configured for that
element.

We start here with the graph elements.

But we leave the ElementConfig a regular dict for now until all report
elements have been upgraded. This way we can prepare the code step by
step for the final change.

Change-Id: I0d6df1437ebc8e6d6b963af2ec09e50ef57d1075