Frontmatter: Render TOC without extending toc config
We need the headlines and page numbers from the original document when
rendering the TOC. This was previously transported by extending the TOC
element config data structure. Instead we now use the report instance
for that, which is a data structure that is intended to be modified
during rendering of the report.

Change-Id: Ibd166cf147f88bb15bdf37191f4c8c1e6637bd0f