14534 Add unit conversion support for graphs
This werk adds the support for converting the displayed unit of a metric
in a graph.

Currently we only support conversion of degrees from Celsius to
Fahrenheit or Kelvin but this can be extended in the future.

To convert a unit you can use the 'Convert unit to' option in
'Graph Tunings'.

SUP-9297

Change-Id: Ic9f34a7e5949a1a71bb652c1374747a743442d58