Make RuleConditions immutable
We would like to gain some clarity in the `load_...(self, ...)` zoo
under `RulesetCollection`.  This is one step in that direction.

Moreover, immutability makes the code safer and easier to understand
and, in the present case at least, greatly simplifies the users.

The set-and-modify pattern used in the callers is verbose and confusing.

CMK-11325

Change-Id: I2247193b6385397d9010b438b259478f92a31303