14863 Timesyncd check plugin: Change time since last sync check behavior
This werk removes the default threshold for the allowed duration since last synchronization of the systemd timesyncd.
The time since last synchronization reported by the modification time of /var/lib/systemd/timesync/clock can at times be unsuitable to monitor when synchronization occurred. Instead of applying thresholds per default, the time since last synchronization is now available as a metric to support monitoring when a time offset occurs.

If you want to apply thresholds to the time since last synchronization you now have to enable the rule manually.

Additionally, the rendering of the jitter was corrected to show a time span instead of a date.

CMK-11089

Change-Id: Ieb4caa96dc38e92033f30a336bd7dbac1b7995df