14869 FIX Fix regression in mk_logwatch plugin in Windows
Until now mk_logwatch plugin could not create a directory for batch
files because the directory name as a rule contained a colon and the
colon is a forbidden symbol in NTFS. Due to this bug the logwatch
monitoring was impossible.

With this version mk_logwatch plugin replaces the colon in directory
name with an underscore thus fixing the regression.

SUP-11644

Change-Id: Ia2362edbf9f1ccb6a54ba97f7046e637d95de426