14509 SEC add authentication to REST API documentation
It was previously not required to be authenticated to access the site's REST API documentation.

Because custom user tags and comments may appear in the automatically generated documentation,
this would represent an "information leak". Therefore, from this Werk onwards, the site's
REST API documentation is only allowed to be accessed by logged in users.

Change-Id: I5efffe69054cff64475be10488ca52e4a85a1ba9