14806 FIX Users created with the REST API can be edited
Previously if a user was created via the REST-API it could not be edited in the GUI. This has been fixed now.

--- Code explanation ---

The decision if a user is editable in the gui is based on the presence
of the "connector" key in the UserSpec, see
`cmk/gui/wato/pages/users.py:362`.

With this change the connector is always set to htpasswd if
authentication for a user is set. If the password is removed the
connector is removed as well.

The choice of htpassword was used, as this is the default observed in a
"heute" site.

Change-Id: I1ea29755d7ec315db170d0ea3808d1b20f6800e0
JIRA-Ref: FEED-7346