15096 Introducing the Checkmk Development Server
This change introduces the GUI development server
which can be run from the command line.

You will need to be in the root of the git repo for
this to work.

Use the following command to get available options
of the server.

$ python -m cmk.gui.dev_server -h

Once it is running, you can log into Checkmk's GUI and
click around. Look at the logs, as they will tell you
where to create your .htpasswd file for your local user(s).

Not supported yet:

 * actually doing monitoring
 * getting status data from CORE
 * sending livestatus commands

For this change to work, the session management of Checkmk
had to be changed to use the one provided by Flask.

Change-Id: Id4d2d781a6161a77921c45669a605ef5fcae965d