14901 FIX user_config: REST API fix to ensure the request and response schemas align
This werk introduces a fix that aligns both the response and request schemas on the user_config
endpoints. The response schema now nests the enforce_password_change attribute just like the
request schemas.  So the response from a GET request has the same format as what is required in
a POST/PUT request. Client scripts should be adjusted accordingly.

Change-Id: Ic7886931391dbb5fe3c28353c7b086d3e8487c7c