14878 kube_collector_info: add option to map missing machine sections and change default state
Previously, the Collector Info service reported an OK State even if the machine sections
were missing. This werk changes this default to a CRIT State if those sections are missing.
In addition, the werk also introduces a service rule which allows the user to map the
monitoring state for such a scenario.

Change-Id: I1de076fa1233c4003f9520dfcf5dec55d057fc2a