14562 FIX Prometheus: allow specifying a path prefix
Previously, the Prometheus special agent would use the path prefix as a
URL base prefix.  For example, specifying the use of the host IP
address, e.g. `1.2.3.4`, together with a prefix, e.g., `my_prefix` would
result in an API call to

```
https://my_prefix1.2.3.4/api/v1/status/buildinfo
```

In particular, this behaviour deviated from the description in the
Prometheus rule. With this werk, the path prefix is prepended to the API
calls instead, e.g.,

```
https://1.2.3.4/my_prefix/api/v1/status/buildinfo
```

Users, which have specified a path prefix, are migrated to the new
option `Custom URL base prefix`, which has the same behaviour as the old
path prefix option.

CMK-11288

Change-Id: I7238475d36cc899e945c9676784c11dc4ee86930