14823 FIX Correction to werk #14717: Compatibility of local GUI extensions
In werk #14717, we introduced an incompatibility with respect to local
GUI extensions. This was not intended, since we try to avoid
incompatibilities in patch releases of stable versions as much as
possible. Hence, with this werk, the incompatiblity is removed.

However, note that this is does not affect the incompatibilites
announced in werk #14716 for the next major release.

Change-Id: I734414c04a56a2ebb15d8a4dc606b958ddd0e1f5