14625 FIX fortigate_sync_status: No longer crashes if the status is not available or is not known
The check used to crash if the status of a cluster was not available (i.e. empty string) or
if the status was not known (wasn't either 'synchronized' or 'unsynchronized').
Now, if any of the above cases occur, the check turns to an UNKNOWN state and provides the appropriate description.

Change-Id: If9451397b34b363c90bafb5fc4e02fdab3309232