Agent controller: Remove receiver port from connection id
The port is now stored in the connection struct used for standard push
and pull connections. This renders existing configurations on disk
invalid. We will add a corresponding update step in a later commit.

Furthermore, we will also re-add the receiver port to the output of the
status command in a later commit.

CMK-11426

Change-Id: I3e1593786c7c2f3de91b168345e9f1701f18d19b