14917 FIX Fix WebUI form submission
With Werk 13903 a new CSRF token was introduced.
This token is added to all AJAX calls so the backend can validate this.
Unfortunately it was forgotten to add this token to the Mobile UI as well.
The missing token raised an error when submitting forms in the WebUI.

The token was added to the mobile UI.

Change-Id: I7eeddaa4f4d7c3c8eb9a9f0b7d25b25062ed6521