14684 FIX Fixed handling of Livestatus socket specification in livedump/cmc
There were a few tiny bugs plus some surprising behavior when specifying the
Livestatus socket for livedump/cmcdump.  With this change, the rules are
simple and consistent now: If there is a -s/--socket commandline option, use
that.  Otherwise derive the UNIX socket from the OMD_ROOT environment
variable if that is set (probably the common case).  And finally: If there
is no -s/--socket option and no OMD_ROOT, complain and exit.

Change-Id: I1eabde2ca1afaeebe83013070b4da1b46dd1fa54