Fix broken cycles test
Seems the command line API has changed and we did not update
our call site. Just unbreak the job, not trying to improve the
robustness of the call site.

Sadly we have moved from 13 to 14 cycles in the meantime. We
will have to figure out how to clean that up again next.

Change-Id: Icc67d879692b6894cdf4706bc72281285a9dd576