Re-organize checking module
Split into three different modules.

The modules are entirely procedural, having a less entry points
clarifies the API.

Having a lot of code in `__init__.py` really is not a good idea in
Python, due to the way modules are imported.

This improves the cohesion withing the modules:  The commandline part
and the active part were not touching each other, but both were calling
into the other functions.  The new structure reflects this.

This further makes it clear that `active` is actually only used in
CEE.  We may want to move this and fix the licencing in another step.

Change-Id: I5f0423aeaaeee250a1f0b23998a4916a7163163e